The problem with this test was that the `before` method only returned values from the `R` compartments.
So at the level of the `testEquivalence` method, we just had for `before` the values of the `R` compartments, on the other hand for the `after` method we have the values of the `S`, `E`, `I` and `R`.

This problem has been solved. So the equivalence test returns the same number of items for both the `before` and `after` method.

One last problem remains, `before` and `after` do not return the same values for `E`, `I` and `R`.